backend/requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
Flask-Cors
Flask-Marshmallow
marshmallow-sqlalchemy
python-dotenv